Update handling of JSON URL and popup template params
Change the name of point data URL option from ``points`` to
``pointsUrl`` and expose it via data attributes.

Change the popup template option name from ``template`` to
``popupTemplate``.

Support either a Mustache template for the popup template or a
string-returning function.

Factor out the adding of a point layer from a GeoJSON URL.